Описание моделей перенесено в отдельный файл. Обновлено README.md
