+++ README.md
@@ -1,13 +1,13 @@
 ## Привет!
 
-### Это проект домашнего задания №18 SkyPro. Простая версия
+### Это проект домашнего задания №18 SkyPro. Сложная версия (MVC архитектура)
 
-В данном домашнем задании было создано небольшое Flask приложение с построением архитектуры, использованием flask-restx
-и базы данных SQLite.
+В данном домашнем задании было создано небольшое Flask приложение с построением MVC архитектуры, 
+использованием flask-restx и базы данных SQLite.
 
 
 **Отработаны навыки** работы с SQLAlchemy, работы с Class-Based-View(CBV) фреймворка flask-restx, написание и
-использование сериализаторов фреймворка marshmallow, а также навыки построения архитектуры Flask приложения
+использование сериализаторов фреймворка marshmallow, а также навыки построения MVC архитектуры Flask приложения.
 
 Для работы необходимо клонировать репозиторий и установить flask, flask-sqlalchemy, sqlalchemy, marshmallow,
 flask_restx. Для этого используйте файл requrements.txt (pip/pip3 install -r requirements.txt)
