+++ dao/models/director.py
@@ -1,0 +1,21 @@
+from marshmallow import Schema, fields
+
+from setup_db import db
+
+
+class Director(db.Model):
+    """
+    Сущность SQLAlchemy, описывающая таблицу с режиссерами
+    """
+    __tablename__ = 'director'
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(255))
+
+
+class DirectorSchema(Schema):
+    """
+    Объект пакета marshmallow
+    Схема для сериализации/десериализации данных о режиссерах
+    """
+    id = fields.Int()
+    name = fields.String()
